Move draw_fuel_alarm_gl to gl::render_gl for better organization
